# shop-service/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Mağazaların CRUD idarəetməsi + mağaza açma müraciəti/təsdiq axını. Birbaşa yaratma yalnız administrator üçündür; adi istifadəçilər /shop-applications ilə müraciət edir, administrator formu göndərəndə (send-form) müvəqqəti mağaza yaranır və müraciət sahibi onun admin(4) səviyyəli sahibi olur, yekun approve ilə mağaza daimi olur.",
        "title": "Shop Service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/shop-applications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop-applications"
                ],
                "summary": "List shop applications (administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pending | form_sent | approved | rejected",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/shop-service_internal_models.ShopApplication"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "İstənilən login olmuş istifadəçi müraciət göndərə bilər. Müraciət \"pending\" statusunda yaranır — administrator ya birbaşa rədd edir, ya da formu göndərir (/send-form), bundan sonra mağaza açılır.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop-applications"
                ],
                "summary": "Apply to open a shop",
                "parameters": [
                    {
                        "description": "Application payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.applicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/shop-service_internal_models.ShopApplication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shop-applications/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Administrator, ya da müraciəti göndərən istifadəçinin özü baxa bilər.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop-applications"
                ],
                "summary": "Get a shop application",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shop-service_internal_models.ShopApplication"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shop-applications/{id}/approve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Yalnız \"form_sent\" statusundan çağırıla bilər. Müvəqqəti mağazanı daimi edir (public siyahılarda görünməyə başlayır).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop-applications"
                ],
                "summary": "Approve a shop application (administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shop-service_internal_models.ShopApplication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shop-applications/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"pending\" statusundan (heç bir mağaza yaranmayıb) və ya \"form_sent\" statusundan (müvəqqəti mağaza silinir, təyinat geri alınır) çağırıla bilər.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop-applications"
                ],
                "summary": "Reject a shop application (administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reject payload",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.rejectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shop-service_internal_models.ShopApplication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/shop-applications/{id}/send-form": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Müraciəti irəli aparmaq qərarı: müvəqqəti mağaza yaradır, müraciət sahibini onun admin(4) səviyyəsinə təyin edir ki, mağaza məlumatlarını özü doldursun (PUT /shops/{id}). Mağaza yekun təsdiqə qədər (approve) public siyahılarda görünmür.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shop-applications"
                ],
                "summary": "Send the shop form (administrator only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/shop-service_internal_models.ShopApplication"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handlers.applicationRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_handlers.rejectRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "shop-service_internal_models.ShopApplication": {
            "type": "object",
            "properties": {
                "applicant_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "review_note": {
                    "type": "string"
                },
                "reviewer_id": {
                    "type": "string"
                },
                "shop_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "template_version": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}